requirements.txt
numpy<2.0,>=1.24.0
opencv-python>=4.8.0,<5.0.0
opencv-python-headless>=4.9.0.80
scikit-learn
scipy
insightface
Pillow
pandas
Flask>=2.3.0
flask-cors>=4.0.0
